progress fixing more warnings

+++ modules/mrpt_maps/include/mrpt/maps/CRandomFieldGridMap2D.h
@@ -315,7 +315,7 @@ class CRandomFieldGridMap2D :
     /** image name of the occupancy_gridmap */
     std::string GMRF_gridmap_image_file;
     /** occupancy_gridmap resolution: size of each pixel (m) */
-    double GMRF_gridmap_image_res{0.01f};
+    float GMRF_gridmap_image_res{0.01f};
     /** Pixel coordinates of the origin for the occupancy_gridmap */
     size_t GMRF_gridmap_image_cx{0};
     /** Pixel coordinates of the origin for the occupancy_gridmap */
@@ -478,7 +478,10 @@ class CRandomFieldGridMap2D :
    * variances are up-to-date with all inserted observations. */
   void updateMapEstimation();
 
-  void enableVerbose([[maybe_unused]] bool enable_verbose) { this->setMinLoggingLevel(mrpt::system::LVL_DEBUG); }
+  void enableVerbose([[maybe_unused]] bool enable_verbose)
+  {
+    this->setMinLoggingLevel(mrpt::system::LVL_DEBUG);
+  }
   bool isEnabledVerbose() const { return this->getMinLoggingLevel() == mrpt::system::LVL_DEBUG; }
 
   void enableProfiler(bool enable = true) { this->m_gmrf.enableProfiler(enable); }
